rearrange encoder/decoder/mae
